Quantify Enums over Set, get rid of type/tdenote
in exchange we depend on an axiom: UIP aka the K axiom.

The alternative would be to only quantify Enum over Sets with decidable
equality, but the mechanics of doing this would clutter the development
for what is ultimately not that important.

On the other hand it might be possible to go without UIP/K/JMeq entirely
using similar techniques described at the end of
http://adam.chlipala.net/cpdt/html/Universes.html but my Coq Fu is too
weak.